Function calibrate_pmt_sum take list of active pmts in argument.
Previous version of function assumed that all PMTs were active and
performed the sum over all PMTs. This is not the general case. The new
version of the function takes a list of active PMTs in the argument
and performs the sum over those PMTs. The default behaviour is the
same than before (if no list is specified the sum is taken over all
PMTs).

Tests have been added to check the new feature